backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-chroma==0.1.4
openai==1.51.0
chromadb==0.5.15
pymupdf==1.24.10
beautifulsoup4==4.12.3
httpx==0.27.0
python-multipart==0.0.12
pydantic==2.9.0
slowapi==0.1.9
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
